split write + read path to blockstore
