+++ history/src/block_stores/postgres/mod.rs
@@ -1,5 +1,5 @@
-pub mod postgres_block_store;
-pub mod postgres_query_block_store;
+pub mod postgres_block_store_writer;
+pub mod postgres_block_store_read_access;
 pub use postgres_config::PostgresSessionConfig;
 pub use postgres_session::PostgresSession;
 pub use postgres_session::PostgresWriteSession;
